src/backend/requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
coveralls==3.3.1
Django==4.1.2
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
docopt==0.6.2
flake8==5.0.4
idna==3.4
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.12
